Place log always in output directory
Previously, the log was placed relative to the directory from where the daemon is called.
This commit adds a handler that will from now on always put the file inside the output
dir using relative path of the file that uses logging. This might not be the best way, but
till we create something to handle directory, this is the best option till then.